feat: implement unavailable state (New without data source)
