+++ src/components/layout-panel/bottom-bar/bottom-bar.tsx
@@ -3,9 +3,11 @@ import { type FC } from 'react'
 import classes from './styles/bottom-bar.module.css'
 import { UpdateButton } from './update-button'
 import { useAppSelector } from '@hooks'
+import { getDataSourceId } from '@store/dimensions-selection-slice'
 import { getIsVisualizationLoading } from '@store/loader-slice'
 
 export const BottomBar: FC = () => {
+    const dataSourceId = useAppSelector(getDataSourceId)
     const isVisualizationLoading = useAppSelector(getIsVisualizationLoading)
 
     return (
@@ -13,6 +15,7 @@ export const BottomBar: FC = () => {
             <div
                 className={cx(classes.container, {
                     [classes.loading]: isVisualizationLoading,
+                    [classes.empty]: !dataSourceId,
                 })}
             >
                 <UpdateButton />
