+++ src/components/layout-panel/styles/axes.module.css
@@ -1,3 +1,7 @@
+.container.empty {
+    min-block-size: 60px;
+}
+
 .axes {
     display: grid;
     grid-template-columns: 2fr 1fr;
@@ -13,6 +17,29 @@
 
 .axes.lineList {
     grid-template-areas: 'columns filters';
+}
+
+.expandButton {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 2px;
+    box-shadow: inset 0 4px 8px -6px rgba(40, 60, 85, 0.05),
+        inset 0 -4px 8px -6px rgba(40, 60, 85, 0.05);
+    background-color: var(--colors-white);
+    cursor: pointer;
+    transition: background-color 0.15s ease, color 0.15s ease;
+}
+
+.expandButton:hover {
+    background-color: var(--colors-grey100);
+}
+
+.loadingSkeletons {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 6px;
+    padding: var(--spacers-dp8) 6px;
 }
 
 .resizeHandle {
