+++ src/components/layout-panel/top-bar/top-bar.tsx
@@ -5,9 +5,11 @@ import { OptionsButton } from './options-button'
 import classes from './styles/top-bar.module.css'
 import { VisualizationTypeSelector } from './visualization-type-selector/visualization-type-selector'
 import { useAppSelector } from '@hooks'
+import { getDataSourceId } from '@store/dimensions-selection-slice'
 import { getIsVisualizationLoading } from '@store/loader-slice'
 
 export const TopBar: FC = () => {
+    const dataSourceId = useAppSelector(getDataSourceId)
     const isVisualizationLoading = useAppSelector(getIsVisualizationLoading)
 
     return (
@@ -20,7 +22,7 @@ export const TopBar: FC = () => {
                 <VisualizationTypeSelector />
                 <OptionsButton />
                 <div className={classes.containerSpacer} />
-                <ExpandedLayoutPanelToggler />
+                {dataSourceId && <ExpandedLayoutPanelToggler />}
             </div>
         </div>
     )
